Debug namespace error & refactor all codes
