+++ Engine/Collision/Collision.h
@@ -1,29 +1,28 @@
 #pragma once
-#include "DataType.h"
+
+#include "Math/Vector.h"
+#include "Math/Matrix.h"
+#include "Debugger.h"
 #include "Mathf.h"
-
-//using namespace Engine;
-
 
 #define M_PI       3.14159265358979323846
 
-
-
+using namespace Engine;
 
 template<typename T>
 class AABB
 {
 public:
-	Vector3<T> center;
-	Vector3<T> extent;
+	Engine::Vector3<T> center;
+	Engine::Vector3<T> extent;
 };
 
 
 struct Object
 {
-	Vector3<float>	m_Position;
-	Vector3<float>	m_Velocity;
-	Matrix3<float>	m_Orientation;
+	Engine::Vector3<float>	m_Position;
+	Engine::Vector3<float>	m_Velocity;
+	Engine::Matrix3<float>	m_Orientation;
 
 	AABB<float>					m_BB;
 };
@@ -38,10 +37,10 @@ struct CollisionPair
 float DegreesToRadians(float i_Degrees);
 
 
-float RandInRange(float i_Min, float i_Max);
+//float RandInRange(float i_Min, float i_Max);
 
 
-Vector3<float> GenerateRandomVector(const Vector3<float>& i_Min, const Vector3<float>& i_Max);
+Vector3<float> GenerateRandomVector(const Engine::Vector3<float>& i_Min, const Engine::Vector3<float>& i_Max);
 
 
 bool DetectCrossTimes(float i_Center, float i_Extent, float i_Point, float i_Travel, float& o_tEnter, float& o_tExit);
