--- Engine/DataType/Engine/GameObject.cpp
@@ -1,25 +1,0 @@
-#include "GameObject.h"
-
-using namespace Engine;
-
-#pragma region GameObject
-
-inline GameObject::GameObject() :
-	name(""),
-	position(Vector2<double>::Zero)
-{ }
-
-inline GameObject::GameObject(const string& name, const Vector2<double>& position) :
-	name(name),
-	position(position)
-{ }
-
-inline GameObject::~GameObject()
-{ }
-
-inline SmartPtr<GameObject> GameObject::Create(const string& name, const Vector2<double>& position)
-{
-	return SmartPtr<GameObject>(new GameObject(name, position));
-}
-
-#pragma endregion
