+++ Engine/DataType/Engines/GameObject.cpp
@@ -1,0 +1,31 @@
+//#include "GameObject.h"
+//
+//namespace Engine
+//{
+//
+//
+//#pragma region GameObject
+//
+//inline GameObject::GameObject() :
+//	name(""),
+//	position(Vector2<double>::Zero)
+//{ }
+//
+//inline GameObject::GameObject(const string& name, const Vector2<double>& position) :
+//	name(name),
+//	position(position)
+//{ }
+//
+//inline GameObject::~GameObject()
+//{ }
+//
+//inline SmartPtr<GameObject> GameObject::Create(const string& name, const Vector2<double>& position)
+//{
+//	return SmartPtr<GameObject>(new GameObject(name, position));
+//}
+//
+//#pragma endregion
+//
+//}
+//
+
