+++ Engine/FileSys/FileSys.cpp
@@ -1,18 +1,20 @@
 #include "FileSys.h"
 
+using std::ios;
+using std::ifstream;
+using std::string;
+
 void ReadJSONFile()
 {
 	// Using fstream to get the file pointer in "file"
 	ifstream file("JsonTest.json", ios::in);
 	
-	
-
 	Json::Value actualJson;
 	Json::Reader reader;
 
 	if (file.fail())
 	{
-		Engine::Debugger::DEBUG_PRINT("Read file fail!! \n ");
+		DEBUG_PRINT("Read file fail!! \n ");
 	}
 	else
 	{
