+++ Engine/FileSys/FileSys.h
@@ -3,7 +3,5 @@
 #include "json/json.h"
 #include "Debugger.h"
 
-using namespace std;
-
 void ReadJSONFile();
