+++ Engine/Mathf/Mathf.h
@@ -3,10 +3,9 @@
 #include <cassert>
 #include "./Utility/Singleton.h"
 
-using namespace std;
-using namespace Engine;
 
-
+namespace Engine
+{
 #define MAX_DIFF 0.00001f
 #define USE_EFFICIENCY
 //#define USE_ACCURACY
@@ -27,7 +26,6 @@ inline int RandInRange(int lowerBound, int upperBound);
 inline float RandInRange(float lowerBound, float upperBound);
 
 template<typename T> inline void Swap(T& left, T& right);
-
 
 
 
@@ -52,19 +50,19 @@ template<typename T> inline void Swap(T& left, T& right);
 class Mathf : public Singleton<Mathf>
 {
 private:
-	random_device rd;			/* Seed source for the random number engine, unique in the whole engine scope */
-	mt19937 generator;			/* Random number engine */
-	uniform_real_distribution<float> distf;
-	uniform_int_distribution<int> disti;
+	std::random_device						rd;			/* Seed source for the random number engine, unique in the whole engine scope */
+	std::mt19937							generator;			/* Random number engine */
+	std::uniform_real_distribution<float>	distf;
+	std::uniform_int_distribution<int>		disti;
 
 	inline Mathf() { Initialize(); }
 	inline ~Mathf() {};
 
 	inline void Initialize()
 	{
-		generator = mt19937(rd());
-		distf = uniform_real_distribution<float>(0.0f, nextafter(1.0f, FLT_MAX));
-		disti = uniform_int_distribution<int>(0, INT_MAX);
+		generator = std::mt19937(rd());
+		distf = std::uniform_real_distribution<float>(0.0f, nextafter(1.0f, FLT_MAX));
+		disti = std::uniform_int_distribution<int>(0, INT_MAX);
 	}
 
 	/** @brief Return a random float value within the range [0.0, 1.0], inclusive in both sides */
@@ -237,7 +235,7 @@ inline void Swap(T& left, T& right)
 inline void RandInRangeIntUnitTest()
 {
 	/* Generate random values */
-	vector<int> samples = vector<int>(1000);
+	std::vector<int> samples = std::vector<int>(1000);
 	for (int i = 0; i < 1000; i++)
 	{
 		samples[i] = RandInRange(1, 100);
@@ -264,14 +262,14 @@ inline void RandInRangeIntUnitTest()
 		&& (mean > expectedMean - threshold)
 		&& (mean < expectedMean + threshold);
 
-	Engine::Debugger::DEBUG_PRINT("Is RandInRangeInt() random: %d\n", res);
+	DEBUG_PRINT("Is RandInRangeInt() random: %d\n", res);
 }
 
 
 inline void RandInRangeFloatUnitTest()
 {
 	/* Generate random values */
-	vector<float> samples = vector<float>(1000);
+	std::vector<float> samples = std::vector<float>(1000);
 	for (int i = 0; i < 1000; i++)
 	{
 		samples[i] = RandInRange(1.0f, 100.0f);
@@ -298,7 +296,9 @@ inline void RandInRangeFloatUnitTest()
 		&& (mean > expectedMean - threshold)
 		&& (mean < expectedMean + threshold);
 
-	Engine::Debugger::DEBUG_PRINT("Is RandInRangeFloat() random: %d\n", res);
+	DEBUG_PRINT("Is RandInRangeFloat() random: %d\n", res);
 }
 
 #endif
+
+}//Namespace Engine
